% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatindexbr-package.R
\docType{package}
\name{heatindexbr-package}
\alias{heatindexbr}
\alias{heatindexbr-package}
\title{heatindexbr: Heat Index Mapping for the Brazilian Semiarid Region}
\description{
The \pkg{heatindexbr} package provides pre-computed Heat Index (HI) rasters
for the Brazilian Semiarid region (1,477 municipalities) at four synoptic
hours (00h, 09h, 15h, 21h local time) and tools for spatial extraction.
\subsection{Main functions}{
\itemize{
\item \code{\link{hi_download}}: Download rasters from Zenodo repository
\item \code{\link{hi_municipality}}: Extract HI for municipalities by name or code
\item \code{\link{hi_shape}}: Extract HI for any area (shapefile or sf object)
\item \code{\link{hi_search}}: Search for valid municipality names
\item \code{\link{hi_plot}}: Generate standardized HI maps
}
}

\subsection{Data}{

Rasters are interpolated using External Drift Kriging with altitude
(SRTM) and ERA5-Land temperature as covariates, validated against
INMET automatic weather station data (2025) with LOOCV R² = 0.884
at 15h local time.
}

\subsection{Citation}{

Campos Neto, A. et al. (2026). Heat Index mapping for the Brazilian
Semiarid: spatial interpolation under sparse monitoring networks.
International Journal of Climatology. (in review)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/antcneto/heatindexbr}
  \item \doi{10.5281/zenodo.21049066}
  \item \doi{10.5281/zenodo.20942619}
  \item \url{https://antcneto.github.io/heatindexbr/}
  \item Report bugs at \url{https://github.com/antcneto/heatindexbr/issues}
}

}
\author{
\strong{Maintainer}: Antônio Campos Neto \email{antonio.camposneto9@gmail.com} (\href{https://orcid.org/0009-0007-3844-3869}{ORCID})

}
\keyword{internal}
