% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{hi_available}
\alias{hi_available}
\title{List available Heat Index products}
\usage{
hi_available()
}
\value{
A \code{data.frame} with columns \code{year},
\code{resolution}, \code{hours}, and \code{status}.
}
\description{
Shows which years, resolutions, and hours are currently available.
}
\examples{
hi_available()

}
