% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_shape.R
\name{hi_shape}
\alias{hi_shape}
\title{Extract Heat Index for a custom spatial area}
\usage{
hi_shape(
  shape,
  hour = "15h",
  year = 2025,
  resolution = "annual",
  month = NULL,
  day = NULL,
  stat = "mean",
  return_raster = TRUE,
  cache_dir = NULL
)
}
\arguments{
\item{shape}{An \code{sf} object or path to a shapefile.}

\item{hour}{Character or NULL. Local hour (default \code{"15h"}).}

\item{year}{Integer. Year (default \code{2025}).}

\item{resolution}{Character. Temporal resolution (default \code{"annual"}).}

\item{month}{Integer. Month if \code{resolution = "monthly"}.}

\item{day}{Integer. Day if \code{resolution = "daily"}.}

\item{stat}{Character vector. Statistics (default \code{"mean"}).}

\item{return_raster}{Logical. Return clipped SpatRaster? Default \code{TRUE}.}

\item{cache_dir}{Character. Cache directory.}
}
\value{
A list with \code{stats} and optionally \code{raster}.
}
\description{
Extract Heat Index for a custom spatial area
}
\examples{
\donttest{
# area <- sf::st_read("my_area.shp")
# res  <- hi_shape(area, hour = "15h")
}

}
